Update progressive_prime_numbers_v1.0
In this first version I left a part of the code where it says "time lost" committed. As I'm doing the calculations and inserting the data based on real time, if the algorithm takes a while to finish the loop it doesn't update the "last_time" column and in the future do a recalculation increasing the time between steps.